Updated data(), where(), select(), insert(), update(). More info about changes cann be found in changelog.md
