+++ src/jbreeze/jbreeze_init.php
@@ -1,0 +1,973 @@
+<?php
+
+namespace Jbreeze;
+
+use jbreezeExceptions\ErrorHandler;
+
+use Exception;
+
+class Jbreeze_init
+{
+
+    protected $data;          // Holds the original dataset
+    private $filteredData;  // Holds the filtered dataset
+    private $isUpdate = false; // Tracks if update was called
+    private $isDelete = false; // Tracks if delete was called
+    private $isInsert = false; // Tracks if insert was called
+    private $newValues = [];   // Stores the values for update or insert
+    private $primaryKey = null; // Stores the primary key for insert
+    protected $jsonFilePath;     // Path to the JSON file (for saving)
+    private $existingKeys = []; // Stores the keys of the first dataset record
+    protected $exceptions = [];  // To store all exception messages as they occur
+
+    protected $config = []; // This configuration is sent to ErrorHandler
+
+    /**
+     * Constructor to initialize configuration.
+     * 
+     * @param array $config Configuration options for the class.
+     */
+    public function __construct(array $config = [])
+    {
+        $this->config = array_merge([
+            'structured' => true // Default is structured JSON
+        ], $config);
+    }
+
+    /**
+     * Toggles between structured and unstructured mode dynamically.
+     * 
+     * @param bool $mode True for structured mode, False for unstructured mode.
+     * @return self
+     */
+    public function jb_init_dataStructure(bool $mode)
+    {
+        $this->config['structured'] = $mode;
+        return $this;
+    }
+
+    /**
+     * Loads data from a file or raw JSON string.
+     * 
+     * @param string $input JSON string or path to a file.
+     * @return self
+     * @throws Exception If the input is invalid or JSON decoding fails.
+     */
+    public function jb_init_data(string $input)
+    {
+        $errorHandler = new ErrorHandler($this->config);
+
+        try {
+
+            if (filter_var($input, FILTER_VALIDATE_URL)) {
+
+                // If input is a URL, fetch JSON data
+                $this->data = $this->jb_fetchJsonFromUrl($input);
+
+            } elseif (is_file($input)) {
+                $this->jsonFilePath = $input;
+
+                // Lock the file for shared access
+                $fileHandle = $this->jb_lockFile(LOCK_SH);
+                if (!$fileHandle) {
+                    throw new Exception("FILE|LOCKFAILED");
+                }
+
+                // Read file contents
+                $jsonContent = file_get_contents($input);
+                fclose($fileHandle); // Close file after reading
+
+                // Handle empty file scenario
+                if (trim($jsonContent) === '') {
+                    $this->data = []; // Empty dataset
+                    $this->existingKeys = []; // No keys yet
+                    return $this; // Allow chaining
+                }
+
+                // Decode JSON data
+                $this->data = $this->jb_decodeJson($jsonContent);
+            } else {
+                // If it's a raw JSON string, decode it
+                $this->data = $this->jb_decodeJson($input);
+            }
+
+            // Ensure valid data format
+            if (!is_array($this->data)) {
+                throw new Exception("JSON|INVALID");
+            }
+
+            // If data is empty, set empty keys for future inserts
+            $this->existingKeys = empty($this->data) ? [] : array_keys(reset($this->data));
+
+            // Initially, filteredData is the full dataset
+            $this->filteredData = $this->data;
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $errorHandler->handle($e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Fetches JSON data from a URL.
+     * 
+     * @param string $url The URL to fetch JSON from.
+     * @return string The JSON response.
+     * @throws Exception If the request fails or the response is not valid JSON.
+     */
+    protected function jb_fetchJsonFromUrl(string $url)
+    {
+        try {
+            $response = file_get_contents($url);
+
+            if ($response === false) {
+                throw new Exception("DATA|URL_FETCH_FAILED");
+            }
+
+            $decode = json_decode($response, true);
+
+            // Check if its an array or a single object
+            if (isset($decode[0]) && is_array($decode[0])) {
+
+                return $decode;
+
+            } else {
+
+                return json_decode("[$response]", true);
+
+            }
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return (new ErrorHandler($this->config))->handle($e->getMessage());
+        }
+    }
+
+    /**
+     * .
+     * 
+     * @param string $data The data to decode.
+     * @return array
+     */
+    protected function jb_decodeJson($data)
+    {
+
+        $decode = json_decode($data, true);
+
+        // Check if its an array or a single object
+        if (isset($decode[0]) && is_array($decode[0])) {
+
+            return $decode;
+
+        } else {
+
+            return json_decode("[$data]", true);
+
+        }
+
+
+    }
+
+    /**
+     * Filters the data based on the given parameters.
+     * Supports dot notation for nested values.
+     * 
+     * @param array $parameters Key-value conditions to filter data.
+     * @return self
+     * @throws Exception If no matching data is found.
+     */
+    public function jb_init_where(array $parameters)
+    {
+        try {
+            $this->filteredData = array_filter($this->filteredData, function ($item) use ($parameters) {
+                foreach ($parameters as $key => $condition) {
+                    // Use dot notation to get the nested value
+                    $value = $this->jb_getValueByDotNotation($item, $key);
+
+                    $orConditions = array_map('trim', explode('||', $condition)); // Split by OR operator `||`
+                    $matched = false;
+
+                    foreach ($orConditions as $subCondition) {
+                        // Check for comparison operators in the sub-condition
+                        if (preg_match('/^([<>]=?|=|%)(.+)$/', $subCondition, $matches)) {
+                            $operator = $matches[1];  // Operator like >, <, >=, etc.
+                            $conditionValue = trim($matches[2]);  // Extract condition value
+
+                            // Handle different comparison operators
+                            switch ($operator) {
+                                case '>':
+                                    if ($value > $conditionValue) {
+                                        $matched = true;
+                                    }
+                                    break;
+                                case '<':
+                                    if ($value < $conditionValue) {
+                                        $matched = true;
+                                    }
+                                    break;
+                                case '>=':
+                                    if ($value >= $conditionValue) {
+                                        $matched = true;
+                                    }
+                                    break;
+                                case '<=':
+                                    if ($value <= $conditionValue) {
+                                        $matched = true;
+                                    }
+                                    break;
+                                case '=':
+                                    if ($value == $conditionValue) {
+                                        $matched = true;
+                                    }
+                                    break;
+                                case '%':  // LIKE condition
+                                    if (stripos($value, $conditionValue) !== false) {
+                                        $matched = true;
+                                    }
+                                    break;
+                            }
+                        } else {
+                            // Default to equality comparison if no operator is present
+                            if ($value == $subCondition) {
+                                $matched = true;
+                            }
+                        }
+
+                        // If one condition matches, stop checking
+                        if ($matched) {
+                            break;
+                        }
+                    }
+
+                    if (!$matched) {
+                        return false;
+                    }
+                }
+                return true;
+            });
+
+            if (empty($this->filteredData)) {
+                throw new Exception("QUERY|NODATAFOUND");
+            }
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Orders the filtered data by a specified column and direction.
+     * 
+     * @param string $column The column to sort by.
+     * @param string $direction Sort direction ('ASC' or 'DESC'). Default is 'DESC'.
+     * @return self
+     * @throws Exception If the column doesn't exist or if data is missing.
+     */
+    public function jb_init_order(string $column, string $direction = 'DESC')
+    {
+        $errorHandler = new ErrorHandler($this->config);
+
+        try {
+            if (empty($this->filteredData)) {
+                throw new Exception("ORDER|NODATA");
+            }
+
+            // Check if column exists in all records
+            foreach ($this->filteredData as $record) {
+                if (!array_key_exists($column, $record)) {
+                    throw new Exception("ORDER|INVALIDCOLUMN: " . $column);
+                }
+            }
+
+            // Sort the data
+            usort($this->filteredData, function ($a, $b) use ($column, $direction) {
+                return strtoupper($direction) === 'ASC'
+                    ? $a[$column] <=> $b[$column]
+                    : $b[$column] <=> $a[$column];
+            });
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $errorHandler->handle($e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Filters data where a specific key's value falls within the given range.
+     * 
+     * @param string $key The key to apply the range filter to.
+     * @param array $range An array containing two values (start and end).
+     * @return self
+     * @throws Exception If the range is invalid or the key is not found.
+     */
+    public function jb_init_between(string $key, array $range = [])
+    {
+        try {
+            // Ensure that the range contains exactly two values
+            if (count($range) !== 2) {
+                throw new Exception("BETWEEN|INVALIDRANGE");
+            }
+
+            // Extract the start and end of the range
+            [$start, $end] = $range;
+
+            // Filter the data to only include records where the key is between the start and end values
+            $this->filteredData = array_filter($this->filteredData, function ($item) use ($key, $start, $end) {
+                if (!isset($item[$key])) {
+                    throw new Exception("BETWEEN|INVALIDKEY: " . $key);
+                }
+
+                return $item[$key] >= $start && $item[$key] <= $end;
+            });
+
+            if (empty($this->filteredData)) {
+                throw new Exception("BETWEEN|NOTFOUND");
+            }
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Selects specific keys from the filtered data.
+     * 
+     * @param array $keys An array of keys to include in the result.
+     * @return self
+     * @throws Exception If any key is not found in the data.
+     */
+    public function jb_init_select(array $keys = [])
+    {
+        try {
+            if (!empty($keys)) {
+                $this->filteredData = array_map(function ($item) use ($keys) {
+                    $selected = [];
+                    foreach ($keys as $key) {
+                        $value = $this->jb_getValueByDotNotation($item, $key);  // Access nested values
+                        if ($value !== null) {
+                            $selected[$key] = $value;
+                        }
+                    }
+                    return $selected;
+                }, $this->filteredData);
+            }
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Retrieves a value from a nested array using dot notation.
+     * 
+     * @param array $item The data array to search.
+     * @param string $key The dot-notated key to retrieve the value.
+     * @return mixed The value at the specified key, or null if not found.
+     */
+    protected function jb_getValueByDotNotation(array $item, string $key)
+    {
+        $keys = explode('.', $key);  // Split the key by dots (e.g., "town.town3" becomes ["town", "town3"])
+
+        foreach ($keys as $innerKey) {
+            if (isset($item[$innerKey])) {
+                $item = $item[$innerKey];  // Filter down into the nested structure
+            } else {
+                return null;  // Return null if any key in the chain is not found
+            }
+        }
+
+        return $item;  // Return the final value
+    }
+
+    /**
+     * Orders the keys of a dataset alphabetically.
+     * Moves the 'id' key to the front if it exists.
+     * 
+     * @param array $data The dataset to be ordered.
+     * @return array The dataset with keys ordered alphabetically.
+     */
+    protected function jb_orderKeysAlphabetically(array $data)
+    {
+        // If $data is an array of records (multidimensional), apply sorting to each record
+        if (is_array($data) && isset($data[0]) && is_array($data[0])) {
+            // Loop through each record and sort the keys
+            return array_map(function ($item) {
+                return $this->jb_orderKeysAlphabetically($item);  // Recursively order each record's keys
+            }, $data);
+        }
+
+        // Sort the keys of a single associative array
+        ksort($data);
+
+        // Check if the "id" key exists, and if so, move it to the front
+        if (isset($data['id'])) {
+            $idValue = $data['id'];
+            unset($data['id']);  // Remove "id" key from its current position
+            $data = ['id' => $idValue] + $data;  // Reinsert "id" at the beginning
+        }
+
+        return $data;
+    }
+
+    /**
+     * Finds and filters the dataset for a specific key-value pair.
+     * 
+     * @param string $key The key to search for.
+     * @param mixed $value The value to match.
+     * @return self
+     * @throws Exception If no matching data is found.
+     */
+    public function jb_init_find(string $key, $value)
+    {
+        try {
+            $this->filteredData = array_filter($this->data, fn($item) => isset($item[$key]) && $item[$key] === $value);
+
+            if (empty($this->filteredData)) {
+                throw new Exception("QUERY|NODATAFOUND");
+            }
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+        }
+
+        return $this;
+    }
+
+    /**
+     * Marks data for update with new values.
+     * Actual update happens when run() is called.
+     * 
+     * @param array $newValues The new values to update.
+     * @return self
+     */
+    public function jb_init_update(array $newValues)
+    {
+        $this->isUpdate = true;
+        $this->newValues = $this->convertDotNotationToNestedUpdate($newValues);
+        return $this;
+    }
+
+    /**
+     * Converts dot notation keys into nested updates while preserving existing data.
+     *
+     * @param array $updates The update data with dot notation.
+     * @return array The transformed nested array with merged updates.
+     */
+    protected function convertDotNotationToNestedUpdate(array $updates)
+    {
+        $nestedUpdates = [];
+
+        foreach ($updates as $key => $value) {
+            if (strpos($key, '.') !== false) {
+                $keys = explode('.', $key);
+                $temp = &$nestedUpdates;
+
+                // Traverse and build the nested array dynamically
+                foreach ($keys as $innerKey) {
+                    if (!isset($temp[$innerKey]) || !is_array($temp[$innerKey])) {
+                        $temp[$innerKey] = [];
+                    }
+                    $temp = &$temp[$innerKey]; // Move reference deeper
+                }
+
+                // Assign the updated value to the final nested key
+                $temp = $value;
+            } else {
+                // If no dot notation, assign value normally
+                $nestedUpdates[$key] = $value;
+            }
+        }
+
+        return $nestedUpdates;
+    }
+
+    /**
+     * Marks data for insertion with new values.
+     * 
+     * @param array $newValues The new record values.
+     * @param string|null $primaryKey The primary key field to auto-increment (optional).
+     * @return self
+     * @throws Exception If the insert operation fails (e.g., invalid keys).
+     */
+    public function jb_init_insert(array $newValues, ?string $primaryKey = null)
+    {
+        $errorHandler = new ErrorHandler($this->config);
+
+        try {
+            $this->isInsert = true;
+            $this->newValues = $this->convertDotNotationToNested($newValues); // Convert dot notation keys
+            $this->primaryKey = $primaryKey;
+
+            // Check if dataset is empty
+            if (empty($this->data)) {
+                
+                // If a primary key is specified, ensure it exists in the new values
+                if ($primaryKey && !isset($newValues[$primaryKey])) {
+                    $this->newValues[$primaryKey] = 1; // Start primary key from 1
+                }
+
+                $this->data[] = $this->newValues;
+
+                // Save immediately if a JSON file exists
+                if ($this->jsonFilePath) {
+                    $this->jb_saveToFile();
+                }
+
+                return $this;
+            }
+
+            // If there are existing records, validate the keys
+            $lastRecord = end($this->data);
+            $existingKeys = array_keys($lastRecord);
+
+            // Validate new structure if structureed mode is enabled
+            if ($this->config['structured']) {
+                $this->jb_validateNewRecordKeys($newValues, $existingKeys);
+            }
+
+            // Assign a new primary key value if applicable
+            if ($primaryKey) {
+                if (!$this->jb_validatePrimaryKey($primaryKey, $existingKeys)) {
+                    throw new Exception("KEY|INVALID");
+                }
+
+                // Get the next primary key only if there is existing data
+                $this->newValues[$primaryKey] = $this->jb_getNextPrimaryKeyValue($primaryKey);
+            }
+
+            // Order keys alphabetically
+            $this->newValues = $this->jb_orderKeysAlphabetically($this->newValues);
+
+            // Add new record to data
+            $this->data[] = $this->newValues;
+
+            // If JSON file exists, save immediately
+            // if ($this->jsonFilePath) {
+            //     $this->jb_saveToFile();
+            // }
+
+            return $this;
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $errorHandler->handle($e->getMessage());
+        }
+    }
+
+    /**
+     * Converts dot notation keys (e.g., 'user.id' => 2) into a nested array structure.
+     *
+     * @param array $data The flat associative array with dot notation keys.
+     * @return array The transformed nested array.
+     */
+    protected function convertDotNotationToNested(array $data)
+    {
+        $nestedData = [];
+
+        foreach ($data as $key => $value) {
+            // Check if the key uses dot notation
+            if (strpos($key, '.') !== false) {
+                $keys = explode('.', $key);
+                $temp = &$nestedData;
+
+                // Traverse and build the nested array dynamically
+                foreach ($keys as $innerKey) {
+                    if (!isset($temp[$innerKey]) || !is_array($temp[$innerKey])) {
+                        $temp[$innerKey] = [];
+                    }
+                    $temp = &$temp[$innerKey]; // Move reference deeper
+                }
+
+                // Assign value to the deepest nested key
+                $temp = $value;
+            } else {
+                // If no dot notation, assign value normally
+                $nestedData[$key] = $value;
+            }
+        }
+
+        return $nestedData;
+    }
+
+    /**
+     * Marks the filtered data for deletion.
+     * Actual deletion happens when run() is called.
+     * 
+     * @return self
+     */
+    public function jb_init_delete()
+    {
+        $this->isDelete = true;
+        return $this;
+    }
+
+    /**
+     * Limits the number of results returned from the filtered data.
+     * 
+     * @param int $count The number of results to return.
+     * @return self
+     */
+    public function jb_init_limit(int $count)
+    {
+        try {
+            $this->filteredData = array_slice($this->filteredData, 0, $count);
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+        }
+        return $this;
+    }
+
+    /**
+     * Returns the count of filtered results.
+     * 
+     * @return int The number of filtered records.
+     */
+    public function jb_init_count()
+    {
+        return count($this->filteredData);
+    }
+
+    /**
+     * Retrieves the error log from the ErrorHandler.
+     * 
+     * @return array The array of logged errors.
+     */
+    public static function jb_init_errorslog()
+    {
+
+        $errorHandler = new ErrorHandler();
+
+        return $errorHandler->GetErrorsLog();
+
+    }
+
+    /**
+     * Executes the query and performs any pending insert, update, or delete operations.
+     * Returns the filtered dataset in the specified format.
+     * 
+     * @param string $returnType The format of the return data ('json' or 'array'). Default is 'json'.
+     * @return mixed The filtered dataset in the specified format.
+     * @throws Exception If an error occurs during execution.
+     */
+    public function jb_init_run(string $returnType = 'json')
+    {
+        $this->config['returnType'] = $returnType;
+        $errorHandler = new ErrorHandler($this->config);
+
+        try {
+            if (!empty($this->exceptions)) {
+                // Return an error response using ErrorHandler
+                return $errorHandler->handle($this->exceptions);
+            }
+
+            if ($this->isInsert) {
+                // $this->data[] = $this->newValues;
+                $this->jb_resetFlags();
+                $this->jb_saveToFile();
+                // return $this->jb_finalize();
+            }
+
+            if ($this->isUpdate) {
+                if (empty($this->filteredData)) {
+                    throw new Exception("UPDATE|NOTFOUND");
+                }
+
+                foreach ($this->data as &$item) {
+                    foreach ($this->filteredData as $filteredItem) {
+                        if ($item == $filteredItem) {
+                            $item = array_replace_recursive($item, $this->newValues);
+                        }
+                    }
+                }
+                $this->jb_resetFlags();
+                return $this->jb_finalize();
+            }
+
+            if ($this->isDelete) {
+                if (empty($this->filteredData)) {
+                    throw new Exception("DELETE|NOTFOUND");
+                }
+
+                $originalCount = count($this->data);
+                $this->data = array_filter($this->data, function ($item) {
+                    return !in_array($item, $this->filteredData);
+                });
+                $newCount = count($this->data);
+
+                $this->jb_resetFlags();
+                return $this->jb_finalize($newCount < $originalCount);
+            }
+
+            // Ensure filtered data is not empty before returning results
+            if (empty($this->filteredData)) {
+                throw new Exception("QUERY|NODATAFOUND");
+            }
+
+            $response = [
+                'status' => 'success',
+                'result' => $this->filteredData,
+                'timestamp' => date('c')
+            ];
+
+            $this->jb_resetFlags();
+
+            return $returnType === 'array' ? $response : json_encode($response, JSON_PRETTY_PRINT);
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $errorHandler->handle($this->exceptions);
+        }
+    }
+
+    /**
+     * Finalizes the operation by saving the updated dataset to a file or returning the result.
+     * 
+     * @param bool $operationResult True if the operation (insert/update/delete) succeeded.
+     * @return mixed JSON encoded data or true/false based on file save success.
+     */
+    protected function jb_finalize(bool $operationResult = true)
+    {
+        if ($this->jsonFilePath) {
+
+            try {
+                // Lock the file for exclusive access
+                $fileHandle = $this->jb_lockFile(LOCK_EX);
+                if (!$fileHandle) {
+                    throw new Exception("FILE|LOCKFAILED");
+                }
+
+                // Write data using the locked file handle
+                $saveResult = $this->jb_saveToFile($fileHandle);
+
+                // Unlock the file after saving
+                $this->jb_unlockFile($fileHandle);
+
+                return $saveResult ? true : false;
+
+            } catch (Exception $e) {
+                $this->logException($e->getMessage());
+                return false;
+            }
+
+        } else {
+
+            $response = [
+                'status' => 'success',
+                'result' => $this->data,
+                'timestamp' => date('c')
+            ];
+
+            return $operationResult ? json_encode($response, JSON_PRETTY_PRINT) : false;
+        }
+    }
+
+    /**
+     * Saves the current dataset to a JSON file.
+     * 
+     * @return bool True on success, false on failure.
+     * @throws Exception If an error occurs during file saving.
+     */
+    protected function jb_saveToFile($fileHandle = null)
+    {
+        try {
+            if ($this->jsonFilePath) {
+                $jsonData = json_encode($this->data, JSON_PRETTY_PRINT);
+                if ($fileHandle) {
+                    // Truncate the file, rewind the pointer, and write the new content
+                    ftruncate($fileHandle, 0);
+                    rewind($fileHandle);
+                    if (fwrite($fileHandle, $jsonData) === false) {
+                        throw new Exception("FILE|SAVEERROR");
+                    }
+                } else {
+                    // Fallback if no file handle is provided
+                    if (file_put_contents($this->jsonFilePath, $jsonData) === false) {
+                        throw new Exception("FILE|SAVEERROR");
+                    }
+                }
+            }
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Retrieves the next available value for a primary key.
+     * 
+     * @param string $primaryKey The primary key field.
+     * @return int|false The next primary key value, or false if an error occurs.
+     * @throws Exception If the primary key value is invalid.
+     */
+    protected function jb_getNextPrimaryKeyValue(string $primaryKey)
+    {
+        $errorHandler = new ErrorHandler($this->config);
+
+        try {
+            // Get valid integer values only
+            $validKeys = array_filter($this->data, fn($record) => isset($record[$primaryKey]) && is_int($record[$primaryKey]));
+
+            if (empty($validKeys)) {
+                return 1; // Start from 1 if no valid primary keys exist
+            }
+
+            // Get the max valid primary key value
+            $nextKey = max(array_column($validKeys, $primaryKey)) + 1;
+
+            return $nextKey;
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $errorHandler->handle($e->getMessage());
+        }
+    }
+
+    /**
+     * Validates that the primary key exists and has integer values in the dataset.
+     * 
+     * @param string $primaryKey The primary key field.
+     * @param array $existingKeys The existing keys in the dataset.
+     * @return bool True if the primary key is valid, false otherwise.
+     * @throws Exception If the primary key is invalid.
+     */
+    protected function jb_validatePrimaryKey(string $primaryKey, array $existingKeys)
+    {
+        try {
+            // Check if the primary key exists in the dataset keys
+            if (!in_array($primaryKey, $existingKeys)) {
+                throw new Exception("KEY|NOTFOUND");
+            }
+
+            // Validate that the primary key values are integers
+            foreach ($this->data as $record) {
+                if (isset($record[$primaryKey]) && !is_int($record[$primaryKey])) {
+                    throw new Exception("KEY|INVALID");
+                }
+            }
+
+            return true; // Validation successful
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return false; // Validation failed
+        }
+    }
+
+    /**
+     * Validates that new record keys match the existing dataset keys.
+     * Automatically adds missing keys with null values.
+     * 
+     * @param array $newValues The new record values.
+     * @param array $existingKeys The keys from the existing dataset.
+     * @throws Exception If extra keys are found in the new record.
+     */
+    protected function jb_validateNewRecordKeys(array $newValues, array $existingKeys)
+    {
+        try {
+
+            // Only validate key structure when using structured data
+            if ($this->config['structured']) {
+
+                // Extract the keys of the new record
+                $newKeys = array_keys($newValues);
+
+                // Find missing keys in the new record (keys that exist in the last record but not in the new record)
+                $missingKeys = array_diff($existingKeys, $newKeys);
+
+                // Find any extra keys (keys that exist in the new record but not in the last record)
+                $extraKeys = array_diff($newKeys, $existingKeys);
+
+                // Throw exception if there are any extra keys that don't exist in the last record
+                if (!empty($extraKeys)) {
+                    throw new Exception("INSERT|EXTRAKEY");
+                }
+
+                // Automatically add missing keys with empty values (null)
+                foreach ($missingKeys as $missingKey) {
+                    $this->newValues[$missingKey] = '';
+                }
+
+            }
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $this;  // Ensure the method returns $this even after exception
+        }
+    }
+
+    /**
+     * Resets all flags and the filtered data after an operation.
+     * 
+     * @return void
+     */
+    protected function jb_resetFlags()
+    {
+        $this->isUpdate = false;
+        $this->isDelete = false;
+        $this->isInsert = false;
+        $this->primaryKey = null;
+        $this->newValues = [];
+        $this->exceptions = [];
+
+        // Ensure filteredData resets back to the full dataset
+        $this->filteredData = $this->data;
+    }
+
+    /**
+     * Logs an exception message to the internal exceptions array.
+     * 
+     * @param string $message The exception message.
+     * @return void
+     */
+    protected function logException(string $message)
+    {
+        $this->exceptions[] = $message; // Add the exception to the list
+    }
+
+    protected function jb_lockFile($mode, $retryInterval = 100, $timeout = 5000)
+    {
+        $errorHandler = new ErrorHandler($this->config);
+
+        try {
+            if (!$this->jsonFilePath) {
+                throw new Exception("FILE|NOPATH");
+            }
+
+            $fileHandle = fopen($this->jsonFilePath, 'c+');
+            if (!$fileHandle) {
+                throw new Exception("FILE|OPENFAILED");
+            }
+
+            $startTime = microtime(true);
+            while (!flock($fileHandle, $mode)) {
+                usleep($retryInterval * 1000);
+                if ((microtime(true) - $startTime) * 1000 >= $timeout) {
+                    fclose($fileHandle);
+                    throw new Exception("FILE|LOCKTIMEOUT");
+                }
+            }
+
+            return $fileHandle;
+
+        } catch (Exception $e) {
+            $this->logException($e->getMessage());
+            return $errorHandler->handle($e->getMessage());
+        }
+    }
+    protected function jb_unlockFile($fileHandle)
+    {
+        if ($fileHandle) {
+            flock($fileHandle, LOCK_UN);  // Release the lock
+            fclose($fileHandle);         // Close the file handle
+        }
+    }
+}
+
